ML_assignment/requirements.txt
# Core dependencies
python-docx>=1.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# LLM and RAG frameworks
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
langchain-core>=0.1.0
openai>=1.6.0

# Vector database
faiss-cpu>=1.7.4
chromadb>=0.4.18

# Document processing
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0
html5lib>=1.1

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pandas>=2.1.0
numpy>=1.24.0

# Optional: for better HTML parsing
html2text>=2020.1.16

